requirements.txt
torch
torchvision
kagglehub
flask
pillow
grad-cam
requests
